dslp {meteo}	R Documentation

Mean sea level pressure in hPa for July 2011

Description

Sample data set showing values of merged mean sea level pressure measurements from the Global Surface Summary of Day data (GSOD) with European Climate Assessment & Data set (ECA&D) for the month July 2011.

Usage

data(dslp)

Format

The dslp contains the following columns:

staid

character; station ID from GSOD or ECA\&D data set

time

Date; day of the measurement

slp

numeric; mean sea level pressure amount in hPa

Note

The data summaries provided here are based on data exchanged under the World Meteorological Organization (WMO) World Weather Watch Program. To prepare a point map, merge with the stations table containing stations' coordinates.

Author(s)

Milan Kilibarda and Tomislav Hengl

References

Global Surface Summary of the day data (ftp://ftp.ncdc.noaa.gov/pub/data/gsod/)

European Climate Assessment & Dataset (http://eca.knmi.nl/dailydata/predefinedseries.php)

Examples

## load data
data(dslp)
str(dslp)

[Package meteo version 0.1-5 Index]