regdata {meteo}	R Documentation

Dynamic and static covariates for spatio-temporal regression kriging

Description

Dynamic and static covariates for spatio-temporal regression kriging of STFDF-class. The regdata contains geometrical temperature trend, MODIS LST 8-day splined at daily resolution, elevation and topographic wetness index.

Usage

data(regdata)

Format

The regdata contains the following dynamic and static covariates:

regdata$temp_geo

numeric; geometrical temperature trend for mean temperature, calculated with tgeom2STFDF ; from 2011-07-05 to 2011-07-09, in degree Celsius

regdata$modis

numeric; MODIS LST 8-day splined at daily resolution, missing pixels are filtered by spatial splines and 8-day values are splined at daily level; from 2011-07-05 to 2011-07-09, in degree Celsius

regdata@sp$dem

numeric; elevation data obtained from Worldgrids (http://worldgrids.org/)

regdata@sp$twi

numeric; SAGA Topographic Wetness Index (TWI) from Worldgrids (http://worldgrids.org/)

Author(s)

Milan Kilibarda kili@grf.bg.ac.rs

Examples

data(regdata)
str(regdata)
library(sp) # spplot
library(spacetime) # stplot

stplot(regdata[,,'modis']) # plot modis data

spplot(regdata@sp,zcol='twi', col.regions = bpy.colors() ) # plot TWI

spplot(regdata@sp,zcol='dem', col.regions = bpy.colors() ) # plot dem

[Package meteo version 0.1-5 Index]