nlmodis20110704 {meteo}	R Documentation

MODIS LST 8 day images image for the Netherlands ('2011-07-04')

Description

The original 8 day MODIS LST images were also converted from Kelvin to degrees Celsius using the formula indicated in the MODIS user's manual.SpatialGridDataFrame.

Usage

data(nlmodis20110704)

Author(s)

Milan Kilibarda kili@grf.bg.ac.rs

References

Wan, Z., Y. Zhang, Q. Zhang, and Z.-L. Li (2004), Quality assessment and validation of the MODIS global land surface temperature, Int. J. Remote Sens., 25(1), 261-274

Examples

data(nlmodis20110704)
# spplot(nlmodis20110704)

[Package meteo version 0.1-5 Index]