rfillspgaps {meteo}	R Documentation

Close gaps of a grid or raster Layer data

Description

The function close gaps of a grid or raster Layer data by using IDW.

Usage

 rfillspgaps(rasterLayer,maskPol=NULL,nmax =50,zcol=1, ...)

Arguments

rasterLayer
raster Layer or SpatiaGrid(Pixels)DF containing NAs

maskPol
SpatialPolygons or SpatialPolygonsDataFrame

nmax
see idw

zcol
variable column name or number showing position of variable in rasterLayer to be interpolated

...
arguments passed to idw

Value

raster Layer or SpatiaGrid(Pixels)DF object with NA replaced using IDW

Author(s)

Milan Kilibarda kili@grf.bg.ac.rs

References

Kilibarda, M., T. Hengl, G. B. M. Heuvelink, B. Graeler, E. Pebesma, M. Percec Tadic, and B. Bajat (2014), Spatio-temporal interpolation of daily temperatures for global land areas at 1 km resolution, J. Geophys. Res. Atmos., 119, 2294-2313, doi:10.1002/2013JD020803;

Kilibarda M., M. Percec Tadic, T. Hengl, J. Lukovic, B. Bajat - Spatial Statistics (2015), Global geographic and feature space coverage of temperature data in the context of spatio-temporal interpolation, doi:10.1016/j.spasta.2015.04.005.

See Also

rfilltimegaps pred.strk

Examples

   library(raster)
   data(nlmodis20110712)
   data(NLpol)
   
   # spplot(nlmodis20110712, col.regions=bpy.colors())
   # fill spatial gaps
   r=rfillspgaps(nlmodis20110712,NLpol)
   # spplot(r, col.regions=bpy.colors())
  

[Package meteo version 0.1-5 Index]