tregcoef {meteo}	R Documentation

Multiple linear regression coefficients for global daily air temperatures

Description

Multiple linear regression coefficients for mean, min., max. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index. The models is computed from GSOD and ECA&D and GHCN-Daily data.

Usage

data(tregcoef)

Format

A list of 8 multiple linear regression coefficients for daily air temperatures.

tmeanGSODECAD

Multiple linear regression coefficients of mean daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GSOD and ECA&D

tmeanGSODECAD_noMODIS

Multiple linear regression coefficients of mean daily temperature on geometric temperature trend, elevation, and topographic wetness index, data used: GSOD and ECA&D

tminGSODECAD

Multiple linear regression coefficients of min. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GSOD and ECA&D

tminGHCND

Multiple linear regression coefficients of min. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GHCN-Daily

tminGSODECAD_noMODIS

Multiple linear regression coefficients of min. daily temperature on geometric temperature trend, elevation, and topographic wetness index, data used: GSOD and ECA&D

tmaxGSODECAD

Multiple linear regression coefficients of max. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GSOD and ECA&D

tmaxGHCND

Multiple linear regression coefficients of max. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GHCN-Daily

tmaxGSODECAD_noMODIS

Multiple linear regression coefficients of max. daily temperature on geometric temperature trend, elevation, and topographic wetness index, data used: GSOD and ECA&D

Author(s)

Milan Kilibarda

References

Kilibarda, M., T. Hengl, G. B. M. Heuvelink, B. Graeler, E. Pebesma, M. Percec Tadic, and B. Bajat (2014), Spatio-temporal interpolation of daily temperatures for global land areas at 1 km resolution, J. Geophys. Res. Atmos., 119, 2294-2313, doi:10.1002/2013JD020803.

Examples

data(tregcoef)
tregcoef[[1]] # model for mean daily temp.

[Package meteo version 0.1-5 Index]