tvgms {meteo}	R Documentation

Spatio-temporal variogram models for global daily air temperatures

Description

Variograms of residuals from multiple linear regression of mean, min., max. daily temperatures on geometric temperature trend, MODIS LST, elevation, and topographic wetness index. The models is computed from GSOD and ECA&D and GHCN-Daily data. The obtained global models for mean, minimum, and maximum temperature can be used to produce gridded images of daily temperatures at high spatial and temporal resolution.

Usage

data(tvgms)

Format

A list of 8 space-time sum-metric models for daily air temperatures, units: space km, time days.

tmeanGSODECAD

Variogram for residuals from multiple linear regression of mean daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GSOD and ECA&D

tmeanGSODECAD_noMODIS

Variogram for residuals from multiple linear regression of mean daily temperature on geometric temperature trend, elevation, and topographic wetness index, data used: GSOD and ECA&D

tminGSODECAD

Variogram for residuals from multiple linear regression of min. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GSOD and ECA&D

tminGHCND

Variogram for residuals from multiple linear regression of min. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GHCN-Daily

tminGSODECAD_noMODIS

Variogram for residuals from multiple linear regression of min. daily temperature on geometric temperature trend, elevation, and topographic wetness index, data used: GSOD and ECA&D

tmaxGSODECAD

Variogram for residuals from multiple linear regression of max. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GSOD and ECA&D

tmaxGHCND

Variogram for residuals from multiple linear regression of max. daily temperature on geometric temperature trend, MODIS LST, elevation, and topographic wetness index, data used: GHCN-Daily

tmaxGSODECAD_noMODIS

Variogram for residuals from multiple linear regression of max. daily temperature on geometric temperature trend, elevation, and topographic wetness index, data used: GSOD and ECA&D

Author(s)

Milan Kilibarda

References

Kilibarda, M., T. Hengl, G. B. M. Heuvelink, B. Graeler, E. Pebesma, M. Percec Tadic, and B. Bajat (2014), Spatio-temporal interpolation of daily temperatures for global land areas at 1 km resolution, J. Geophys. Res. Atmos., 119, 2294-2313, doi:10.1002/2013JD020803.

Examples

data(tvgms)
tvgms[[1]] # model for mean daily temp.

[Package meteo version 0.1-5 Index]